% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracodecode.R
\name{draco_decode}
\alias{draco_decode}
\title{Decode Draco encoded raw bytes containing mesh or point cloud data}
\usage{
draco_decode(data)
}
\arguments{
\item{data}{\code{\link{raw}} bytes containing Draco data e.g. as read by
\code{\link{readBin}}}
}
\value{
a list containing elements \code{points} and (for meshes)
  \code{faces}.
}
\description{
Decode Draco encoded raw bytes containing mesh or point cloud data
}
\examples{

\dontrun{
## convert to standard rgl::mesh3d object
res=draco_decode(data)
m=rgl::tmesh3d(vertices=res$points, indices=res$faces, homogeneous = FALSE)
}
}
