% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracor-package.R
\docType{package}
\name{dracor-package}
\alias{dracor}
\alias{dracor-package}
\title{dracor: Decode Draco Format 3D Mesh Data}
\description{
Decodes meshes and point cloud data encoded by the Draco mesh
    compression library from Google. Note that this is only designed for basic
    decoding and not intended as a full scale wrapping of the Draco library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferis/dracor}
  \item \url{https://github.com/google/draco}
  \item Report bugs at \url{https://github.com/jefferis/dracor/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

Authors:
\itemize{
  \item Google Inc (for the Draco library) [copyright holder]
}

}
\keyword{internal}
