% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracor-package.R
\docType{package}
\name{dracor-package}
\alias{dracor}
\alias{dracor-package}
\title{dracor: Decode Draco Format 3D Mesh data}
\description{
Decodes meshes encoded by Draco
}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

}
\keyword{internal}
